Refactor (REINDEX, Reindex, reindex) -> (INDEX, Index, index) + smaller scope to components scan
