+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/ElasticsearchService.java
@@ -12,8 +12,8 @@ import org.elasticsearch.common.collect.Iterators;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentFactory;
 import org.molgenis.data.*;
-import org.molgenis.data.elasticsearch.converter.ElasticsearchIndexCreator;
-import org.molgenis.data.elasticsearch.converter.MappingsBuilder;
+import org.molgenis.data.elasticsearch.index.ElasticsearchIndexCreator;
+import org.molgenis.data.elasticsearch.index.MappingsBuilder;
 import org.molgenis.data.elasticsearch.request.SearchRequestGenerator;
 import org.molgenis.data.elasticsearch.response.ResponseParser;
 import org.molgenis.data.elasticsearch.util.ElasticsearchUtils;
