+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/index/MappingsBuilder.java
@@ -1,8 +1,7 @@
-package org.molgenis.data.elasticsearch.converter;
+package org.molgenis.data.elasticsearch.index;
 
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.molgenis.MolgenisFieldTypes.AttributeType;
-import org.molgenis.data.elasticsearch.converter.ElasticsearchIndexCreator;
 import org.molgenis.data.elasticsearch.util.MapperTypeSanitizer;
 import org.molgenis.data.meta.model.AttributeMetaData;
 import org.molgenis.data.meta.model.EntityMetaData;
