+++ molgenis-data-elasticsearch/src/main/java/org/molgenis/data/elasticsearch/index/job/IndexJobExecutionMeta.java
@@ -1,4 +1,4 @@
-package org.molgenis.data.elasticsearch.index.meta;
+package org.molgenis.data.elasticsearch.index.job;
 
 import org.molgenis.data.jobs.model.JobExecutionMetaData;
 import org.molgenis.data.meta.SystemEntityMetaData;
